+++ sdk/search/azure-search-documents/src/main/java/com/azure/search/documents/indexes/SearchIndexerAsyncClient.java
@@ -1062,6 +1062,16 @@ public class SearchIndexerAsyncClient {
     /**
      * Resets specific documents in the datasource to be selectively re-ingested by the indexer.
      *
+     * <!-- src_embed com.azure.search.documents.indexes.SearchIndexerAsyncClient.resetDocuments#String-Boolean-List-List -->
+     * <pre>
+     * &#47;&#47; Reset the documents with keys 1234 and 4321.
+     * searchIndexerAsyncClient.resetDocuments&#40;&quot;searchIndexer&quot;, false, Arrays.asList&#40;&quot;1234&quot;, &quot;4321&quot;&#41;, null&#41;
+     *     &#47;&#47; Clear the previous documents to be reset and replace them with documents 1235 and 5231.
+     *     .then&#40;searchIndexerAsyncClient.resetDocuments&#40;&quot;searchIndexer&quot;, true, Arrays.asList&#40;&quot;1235&quot;, &quot;5321&quot;&#41;, null&#41;&#41;
+     *     .subscribe&#40;&#41;;
+     * </pre>
+     * <!-- end com.azure.search.documents.indexes.SearchIndexerAsyncClient.resetDocuments#String-Boolean-List-List -->
+     *
      * @param indexerName The name of the indexer to reset documents for.
      * @param overwrite If false, keys or IDs will be appended to existing ones. If true, only the keys or IDs in this
      * payload will be queued to be re-ingested.
@@ -1070,7 +1080,7 @@ public class SearchIndexerAsyncClient {
      * @return A response signalling completion.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Void> resetDocuments(String indexerName, boolean overwrite, List<String> documentKeys,
+    public Mono<Void> resetDocuments(String indexerName, Boolean overwrite, List<String> documentKeys,
         List<String> datasourceDocumentIds) {
         return withContext(context -> resetDocumentsWithResponse(indexerName, overwrite, documentKeys,
             datasourceDocumentIds, context))
@@ -1079,6 +1089,25 @@ public class SearchIndexerAsyncClient {
 
     /**
      * Resets specific documents in the datasource to be selectively re-ingested by the indexer.
+     *
+     * <!-- src_embed com.azure.search.documents.indexes.SearchIndexerAsyncClient.resetDocumentsWithResponse#SearchIndexer-Boolean-List-List -->
+     * <pre>
+     * searchIndexerAsyncClient.getIndexer&#40;&quot;searchIndexer&quot;&#41;
+     *     .flatMap&#40;searchIndexer -&gt; searchIndexerAsyncClient.resetDocumentsWithResponse&#40;searchIndexer, false,
+     *         Arrays.asList&#40;&quot;1234&quot;, &quot;4321&quot;&#41;, null&#41;
+     *         .flatMap&#40;resetDocsResult -&gt; &#123;
+     *             System.out.printf&#40;&quot;Requesting documents to be reset completed with status code %d.%n&quot;,
+     *                 resetDocsResult.getStatusCode&#40;&#41;&#41;;
+     *
+     *             &#47;&#47; Clear the previous documents to be reset and replace them with documents 1235 and 5231.
+     *             return searchIndexerAsyncClient.resetDocumentsWithResponse&#40;searchIndexer, true,
+     *                 Arrays.asList&#40;&quot;1235&quot;, &quot;5321&quot;&#41;, null&#41;;
+     *         &#125;&#41;&#41;
+     *     .subscribe&#40;resetDocsResult -&gt;
+     *         System.out.printf&#40;&quot;Overwriting the documents to be reset completed with status code %d.%n&quot;,
+     *             resetDocsResult.getStatusCode&#40;&#41;&#41;&#41;;
+     * </pre>
+     * <!-- end com.azure.search.documents.indexes.SearchIndexerAsyncClient.resetDocumentsWithResponse#SearchIndexer-Boolean-List-List -->
      *
      * @param indexer The indexer to reset documents for.
      * @param overwrite If false, keys or IDs will be appended to existing ones. If true, only the keys or IDs in this
@@ -1089,7 +1118,7 @@ public class SearchIndexerAsyncClient {
      * @throws NullPointerException If {@code indexer} is null.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Response<Void>> resetDocumentsWithResponse(SearchIndexer indexer, boolean overwrite,
+    public Mono<Response<Void>> resetDocumentsWithResponse(SearchIndexer indexer, Boolean overwrite,
         List<String> documentKeys, List<String> datasourceDocumentIds) {
         if (indexer == null) {
             return monoError(logger, new NullPointerException("'indexer' cannot be null."));
@@ -1536,6 +1565,14 @@ public class SearchIndexerAsyncClient {
     /**
      * Resets skills in an existing skillset in an Azure Cognitive Search service.
      *
+     * <!-- src_embed com.azure.search.documents.indexes.SearchIndexerAsyncClient.resetSkills#String-List -->
+     * <pre>
+     * &#47;&#47; Reset the &quot;myOcr&quot; and &quot;myText&quot; skills.
+     * searchIndexerAsyncClient.resetSkills&#40;&quot;searchIndexerSkillset&quot;, Arrays.asList&#40;&quot;myOcr&quot;, &quot;myText&quot;&#41;&#41;
+     *     .subscribe&#40;&#41;;
+     * </pre>
+     * <!-- end com.azure.search.documents.indexes.SearchIndexerAsyncClient.resetSkills#String-List -->
+     *
      * @param skillsetName The name of the skillset to reset.
      * @param skillNames The skills to reset.
      * @return A response signalling completion.
@@ -1548,6 +1585,16 @@ public class SearchIndexerAsyncClient {
 
     /**
      * Resets skills in an existing skillset in an Azure Cognitive Search service.
+     *
+     * <!-- src_embed com.azure.search.documents.indexes.SearchIndexerAsyncClient.resetSkillsWithResponse#SearchIndexerSkillset-List -->
+     * <pre>
+     * searchIndexerAsyncClient.getSkillset&#40;&quot;searchIndexerSkillset&quot;&#41;
+     *     .flatMap&#40;searchIndexerSkillset -&gt; searchIndexerAsyncClient.resetSkillsWithResponse&#40;searchIndexerSkillset,
+     *         Arrays.asList&#40;&quot;myOcr&quot;, &quot;myText&quot;&#41;&#41;&#41;
+     *     .subscribe&#40;resetSkillsResponse -&gt; System.out.printf&#40;&quot;Resetting skills completed with status code %d.%n&quot;,
+     *         resetSkillsResponse.getStatusCode&#40;&#41;&#41;&#41;;
+     * </pre>
+     * <!-- end com.azure.search.documents.indexes.SearchIndexerAsyncClient.resetSkillsWithResponse#SearchIndexerSkillset-List -->
      *
      * @param skillset The skillset to reset.
      * @param skillNames The skills to reset.