+++ sdk/search/azure-search-documents/src/main/java/com/azure/search/documents/indexes/SearchIndexerClient.java
@@ -905,8 +905,15 @@ public class SearchIndexerClient {
     /**
      * Resets specific documents in the datasource to be selectively re-ingested by the indexer.
      *
-     * <!-- src_embed com.azure.search.documents.indexes.SearchIndexerClient.resetDocuments#String-boolean-List-List -->
-     * <!-- end com.azure.search.documents.indexes.SearchIndexerClient.resetDocuments#String-boolean-List-List -->
+     * <!-- src_embed com.azure.search.documents.indexes.SearchIndexerClient.resetDocuments#String-Boolean-List-List -->
+     * <pre>
+     * &#47;&#47; Reset the documents with keys 1234 and 4321.
+     * searchIndexerClient.resetDocuments&#40;&quot;searchIndexer&quot;, false, Arrays.asList&#40;&quot;1234&quot;, &quot;4321&quot;&#41;, null&#41;;
+     *
+     * &#47;&#47; Clear the previous documents to be reset and replace them with documents 1235 and 5231.
+     * searchIndexerClient.resetDocuments&#40;&quot;searchIndexer&quot;, true, Arrays.asList&#40;&quot;1235&quot;, &quot;5321&quot;&#41;, null&#41;;
+     * </pre>
+     * <!-- end com.azure.search.documents.indexes.SearchIndexerClient.resetDocuments#String-Boolean-List-List -->
      *
      * @param indexerName The name of the indexer to reset documents for.
      * @param overwrite If false, keys or IDs will be appended to existing ones. If true, only the keys or IDs in this
@@ -915,7 +922,7 @@ public class SearchIndexerClient {
      * @param datasourceDocumentIds Datasource document identifiers to be reset.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public void resetDocuments(String indexerName, boolean overwrite, List<String> documentKeys,
+    public void resetDocuments(String indexerName, Boolean overwrite, List<String> documentKeys,
         List<String> datasourceDocumentIds) {
         resetDocumentsWithResponse(new SearchIndexer(indexerName), overwrite, documentKeys, datasourceDocumentIds,
             Context.NONE);
@@ -924,8 +931,23 @@ public class SearchIndexerClient {
     /**
      * Resets specific documents in the datasource to be selectively re-ingested by the indexer.
      *
-     * <!-- src_embed com.azure.search.documents.indexes.SearchIndexerClient.resetDocumentsWithResponse#String-boolean-List-List-Context -->
-     * <!-- end com.azure.search.documents.indexes.SearchIndexerClient.resetDocumentsWithResponse#String-boolean-List-List-Context -->
+     * <!-- src_embed com.azure.search.documents.indexes.SearchIndexerClient.resetDocumentsWithResponse#SearchIndexer-Boolean-List-List-Context -->
+     * <pre>
+     * SearchIndexer searchIndexer = searchIndexerClient.getIndexer&#40;&quot;searchIndexer&quot;&#41;;
+     *
+     * &#47;&#47; Reset the documents with keys 1234 and 4321.
+     * Response&lt;Void&gt; resetDocsResult = searchIndexerClient.resetDocumentsWithResponse&#40;searchIndexer, false,
+     *     Arrays.asList&#40;&quot;1234&quot;, &quot;4321&quot;&#41;, null, new Context&#40;key1, value1&#41;&#41;;
+     * System.out.printf&#40;&quot;Requesting documents to be reset completed with status code %d.%n&quot;,
+     *     resetDocsResult.getStatusCode&#40;&#41;&#41;;
+     *
+     * &#47;&#47; Clear the previous documents to be reset and replace them with documents 1235 and 5231.
+     * resetDocsResult = searchIndexerClient.resetDocumentsWithResponse&#40;searchIndexer, true,
+     *     Arrays.asList&#40;&quot;1235&quot;, &quot;5321&quot;&#41;, null, new Context&#40;key1, value1&#41;&#41;;
+     * System.out.printf&#40;&quot;Overwriting the documents to be reset completed with status code %d.%n&quot;,
+     *     resetDocsResult.getStatusCode&#40;&#41;&#41;;
+     * </pre>
+     * <!-- end com.azure.search.documents.indexes.SearchIndexerClient.resetDocumentsWithResponse#SearchIndexer-Boolean-List-List-Context -->
      *
      * @param indexer The indexer to reset documents for.
      * @param overwrite If false, keys or IDs will be appended to existing ones. If true, only the keys or IDs in this
@@ -937,7 +959,7 @@ public class SearchIndexerClient {
      * @throws NullPointerException If {@code indexer} is null.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Response<Void> resetDocumentsWithResponse(SearchIndexer indexer, boolean overwrite,
+    public Response<Void> resetDocumentsWithResponse(SearchIndexer indexer, Boolean overwrite,
         List<String> documentKeys, List<String> datasourceDocumentIds, Context context) {
         return asyncClient.resetDocumentsWithResponse(indexer.getName(), overwrite, documentKeys, datasourceDocumentIds,
             context).block();
@@ -1327,6 +1349,13 @@ public class SearchIndexerClient {
     /**
      * Resets skills in an existing skillset in an Azure Cognitive Search service.
      *
+     * <!-- src_embed com.azure.search.documents.indexes.SearchIndexerClient.resetSkills#String-List -->
+     * <pre>
+     * &#47;&#47; Reset the &quot;myOcr&quot; and &quot;myText&quot; skills.
+     * searchIndexerClient.resetSkills&#40;&quot;searchIndexerSkillset&quot;, Arrays.asList&#40;&quot;myOcr&quot;, &quot;myText&quot;&#41;&#41;;
+     * </pre>
+     * <!-- end com.azure.search.documents.indexes.SearchIndexerClient.resetSkills#String-List -->
+     *
      * @param skillsetName The name of the skillset to reset.
      * @param skillNames The skills to reset.
      */
@@ -1337,6 +1366,17 @@ public class SearchIndexerClient {
 
     /**
      * Resets skills in an existing skillset in an Azure Cognitive Search service.
+     *
+     * <!-- src_embed com.azure.search.documents.indexes.SearchIndexerClient.resetSkillsWithResponse#SearchIndexerSkillset-List-Context -->
+     * <pre>
+     * SearchIndexerSkillset searchIndexerSkillset = searchIndexerClient.getSkillset&#40;&quot;searchIndexerSkillset&quot;&#41;;
+     *
+     * &#47;&#47; Reset the &quot;myOcr&quot; and &quot;myText&quot; skills.
+     * Response&lt;Void&gt; resetSkillsResponse = searchIndexerClient.resetSkillsWithResponse&#40;searchIndexerSkillset,
+     *     Arrays.asList&#40;&quot;myOcr&quot;, &quot;myText&quot;&#41;, new Context&#40;key1, value1&#41;&#41;;
+     * System.out.printf&#40;&quot;Resetting skills completed with status code %d.%n&quot;, resetSkillsResponse.getStatusCode&#40;&#41;&#41;;
+     * </pre>
+     * <!-- end com.azure.search.documents.indexes.SearchIndexerClient.resetSkillsWithResponse#SearchIndexerSkillset-List-Context -->
      *
      * @param skillset The skillset to reset.
      * @param skillNames The skills to reset.